+++ source/RTSCamera.CommandSystem/src/CommandSystemMissionStartingHandler.cs
@@ -1,8 +1,6 @@
 ﻿using MissionLibrary.Controller;
 using MissionSharedLibrary.Controller;
-using RTSCamera.CommandSystem.Config;
 using RTSCamera.CommandSystem.Logic;
-using RTSCamera.CommandSystem.Patch;
 using RTSCamera.CommandSystem.View;
 using System.Collections.Generic;
 using TaleWorlds.MountAndBlade;