+++ source/RTSCamera/src/Patch/Patch_HideoutMissionController.cs
@@ -4,7 +4,9 @@ using RTSCamera.Config;
 using RTSCamera.Logic;
 using SandBox.Missions.MissionLogics.Hideout;
 using System;
+using System.Linq;
 using System.Reflection;
+using TaleWorlds.Core;
 using TaleWorlds.Engine;
 using TaleWorlds.MountAndBlade;
 
@@ -26,6 +28,11 @@ namespace RTSCamera.Patch
                         BindingFlags.Static | BindingFlags.Public),
                     postfix: new HarmonyMethod(typeof(Patch_HideoutMissionController).GetMethod(
                         nameof(Postfix_StartBossFightBattleMode), BindingFlags.Static | BindingFlags.Public)));
+                harmony.Patch(
+                    typeof(HideoutMissionController).GetMethod(nameof(HideoutMissionController.OnAgentRemoved),
+                        BindingFlags.Instance | BindingFlags.Public),
+                    postfix: new HarmonyMethod(typeof(Patch_HideoutMissionController).GetMethod(
+                        nameof(Postfix_OnAgentRemoved), BindingFlags.Static | BindingFlags.Public)));
             }
             catch (Exception e)
             {
@@ -44,10 +51,30 @@ namespace RTSCamera.Patch
             {
                 return;
             }
-            //if (RTSCameraConfig.Get().FastForwardHideout == FastForwardHideout.Always)
-            //{
-            //    rtsCameraLogic.SwitchFreeCameraLogic.FastForwardHideoutNextTick = true;
-            //}
+            if (RTSCameraConfig.Get().FastForwardHideout == FastForwardHideout.Always)
+            {
+                rtsCameraLogic.SwitchFreeCameraLogic.FastForwardHideoutNextTick = true;
+            }
+        }
+
+        public static void Postfix_OnAgentRemoved(
+            HideoutMissionController __instance,
+            Agent affectedAgent,
+            Agent affectorAgent,
+            AgentState agentState,
+            KillingBlow blow)
+        {
+            if (affectedAgent.IsMainAgent)
+            {
+                var formation = __instance.Mission.PlayerTeam.PlayerOrderController.SelectedFormations.FirstOrDefault();
+                if (formation != null)
+                {
+                    if (formation.GetReadonlyMovementOrderReference().OrderEnum == MovementOrder.MovementOrderEnum.Retreat)
+                    {
+                        __instance.Mission.PlayerTeam.PlayerOrderController.SetOrder(OrderType.Charge);
+                    }
+                }
+            }
         }
     }
 }